+++ backend/BAHelper.Common/DTOs/StatisticData/StatisticDataInfo.cs
@@ -10,7 +10,7 @@ namespace BAHelper.Common.DTOs.StatisticData
     public class StatisticDataInfo
     {
         public TopicTag TaskTopic { get; set; }
-        public double TaskQuality { get; set; }
+        public int TaskQuality { get; set; }
         public int TaskCount { get; set; }
     }
 }
